Preproccessing and CSV creation
This is a fully commented version of the preprocessing code for the .tiff directors. It can take samples from multiple directors and their corresponding labels from different locations, as long as the director of one set of samples has one CSV for all of its labels. 1 to 1. This code also allows you to load a CSV of the form it creates and process it to have pixel-level location information concerning its location in the original image/tiff file. (This is reuploaded to remove random error messages)